[REFACTORING] Split IMAPServerQResyncTest
